fix auth: 엑세스 토큰 만료 시간 자료형 수정

+++ src/main/java/com/_ithon/speeksee/global/auth/controller/AuthController.java
@@ -72,7 +72,7 @@ public class AuthController {
 		// 2. 응답 DTO 생성
 		AccessTokenResponseDto response = new AccessTokenResponseDto(
 			loginResponse.getAccessToken(),
-			loginResponse.getExpiresIn()
+			(int)loginResponse.getExpiresIn()
 		);
 
 		// 3. API 응답 래핑
